opt: move code for index constraint data structures in a separate file
The `index_constraints.go` file is getting pretty big. Moving the
lower-level functionality around spans to a separate file.

Code is just moved in this commit (there are no modifications).

Release note: None
